aeneas_aligner-main/requirements.txt
aeneas==1.7.3.0
gradio==5.34.2
pdfminer.six==20250506
python-docx==1.2.0

def format_srt_time(seconds):
    """Format seconds as SRT time: HH:MM:SS,mmm"""
    hours = int(seconds // 3600)
    minutes = int((seconds % 3600) // 60)
    secs = int(seconds % 60)
    millis = int((seconds % 1) * 1000)
    return f"{hours:02}:{minutes:02}:{secs:02},{millis:03}"
